property and units

+++ app/(root)/(routes)/(owner)/properties/all_units/components/cell-action.tsx
@@ -13,14 +13,15 @@ import {
 import { Button } from "@/components/ui/button"
 import { useRouter } from "next/navigation"
 import { useState } from "react"
-import { OrderProps } from "@/types"
+import { OrderProps, TenantsReducerProps } from "@/types"
 import { AlertModal } from "@/components/modals/alert-modal"
-import { useDispatch } from "react-redux"
+import { useDispatch, useSelector } from "react-redux"
 import { removeOrder } from "@/redux/orders/ordersSlice"
 import { PreviewOrder } from "@/components/modals/preview-order-modal"
 import { UnitColumn } from "./column"
 import { removeUnit } from "@/redux/units/unitsSlice"
 import { PreviewUnit } from "@/components/modals/preview-unit"
+import api from "@/actions/api"
 // import { AlertModal } from "@/components/modals/alert-modal"
 // import { deleteCategory } from "@/app/actions/categories"
 
@@ -33,15 +34,22 @@ export const CellAction : React.FC<CellActionProps> = ({data}) => {
     const router = useRouter()
     const dispatch = useDispatch()
 
+    const tenants = useSelector(({tenantsReducer} : TenantsReducerProps) => tenantsReducer).tenants
+    .filter((tenant)=> tenant.unitId === data._id)
+
     const [open, setOpen] = useState(false)
     const [openPreview, setOpenPreview] = useState(false)
     const [loading, setLoading] = useState(false)
 
     const onDelete = async () => {
-        // fetch
-        dispatch(removeUnit(data._id))
+        if (tenants.length > 0) {
+            toast.error("Remove associated tenants first")
+        } else {
+            await api.delete(`deleteUnit?id=${data._id}`,{validateStatus: () => true})
+            dispatch(removeUnit(data._id))
+            toast.success("Unit Deleted")
+        }
         setOpen(false)
-        toast.success("Unit Deleted")
     }
